Tie sidewalk costs to their respective streets
In OSM data, street attributes are not tied to the sidewalks adjacent to
them. E.g. if you have a highway: primary street, the sidewalks do not
know that they are highway: primary sidewalks

I don't like walking on busy streets, which means that we need some way
to know if a sidewalk is on a busy street

Do some rudimentary parenting. For each sidewalk point pair
* Check if nearby streets are parallel to it
* Check if nearby streets are close to it
* Take the closest/most parallel street and call it our parent
* Use parent attributes in path planning

Note that we use point pairs instead of ways, because one sidewalk way
may be split between multiple streets

Parenting can be visualized with "Debug parenting" checkbox

Details
* ClosestWayCalculator now return which segment it found so that we can
  lookup the parent for this point pair
* WayBuckets size decreased to make debug mode initialization bearable
* NodePairCostMultiplier map refactored into generic NodePair -> ? Map
  for use in App.PointPairToWayMap
* Node cost management was looking a little difficult, take the easy way
  out and just fully recalculate every change
* index.js got nice lil' helper function to hook up a checkbox
* pp_benchmark becoming error prone, add to lint script
* Lots moved from App -> map_data for code reuse