Debug visualization of path planner steps
In some cases we want to watch how the path planner behaves as it
progresses. Change finish path button to set the path end. In debug mode
we will no longer run the path planner, it has to be polled externally

If debug path planning is disabled it will behave almost as before