Initial transit integration
Introduce a rudimentary path planner that attempts to use the transit
data in order to plan a path according to time instead of distance.

Our path planner is based around A*, the previous heuristic for the A*
search was just a straight line distance. This is problematic for a
couple reasons with the introduction of busses

1. Transit works in units of time
2. Sometimes going farther to get on a fast bus is better than going
   towards the destination

Adjust the path planner to...
* Work based off time (convert distance to time by dividing by movement
  speed)
* Move towards the closer of all unhit bus routes and the final
  destination

The thought here is that the bus network is pretty small (in vancouver,
5k stops copmared to 600k osm nodes). If we over-investigate the bus
network it should end up quite fast overall

The path planner now needs to handle transit and walking/driving/biking
differently. The transit stops use a schedule to determine cost instead
of distance. We now return all trip neighbors separately from OSM
neighbors and handle those connections differently

The current iteration is really stupid on calculating the heuristic
score. We iterate all 5k bus stops on each neighbor update. This will be
improved in the future, but it's in our best interests to get what we
have merged. Hide transit integration behind a checkbox to limit the
damage

UI had to get a few new changes. We now have to specify when we start
walking to determine which busses are available, and our movement speed
to figure out when we'll get to the bus stops

Path planner benchmark has been updated to reflect the above changes as
well

This change has been festering for a week or so, so it's "review"
process has been more about long-ish term testing vs closely
investigating the changes. If you see something you don't like in here,
it's likely that it's actually bad.