Transit heuristic improvements
The initial transit integration used a very expensive heuristic which
calculated the distance from each node to all un-hit bus stops. We can
bring the cost down a lot

Re-frame the heuristic as an approximation of the distance to the
closest bus stop. Sample distances in a grid, and bilinearly
interpolate. As long as the sampling is close enough this seems to work
well enough

Many things were tried for what to interpolation, but what ended up
working best was
* Store the vector for each sample to the closest bus stop, still
  computed on CPU the stupid slow way
* Dot product sample -> point onto sample -> stop
* Interpolate these dot products

The intuition here is pretty hazy, but imagine the samples as a vector
field where we are just trying to follow the flow

We implemented as an OpenGL shader first, to visually see how the
algorithm was performing, then ported to CPU and lazily compute samples
as needed. There is no tying of visualization to real performance which
is a bit of a problem, but leaving the visualization in seems better
than deleting it

Visualization uses a texture to sample positions/vectors, so some gui
callbacks were added to support this. In the process we discovered that
generating mipmaps was not required so removed it

Alpha blending needs to be enabled in webgl to support our heuristic
debug overlay